requirements.txt
motor==3.3.2
nextcord==2.6.0
python-dotenv==1.0.1
vkbottle~=4.3.12
loguru~=0.7.2
psutil~=5.9.8
pytz~=2024.1
APScheduler~=3.10.4
python-dateutil~=2.9.0.post0
pymongo~=4.7.3